UserService/requirements.txt
fastapi
uvicorn 
pydantic 
python-jose
bcrypt
passlib
sqlalchemy
psycopg2
psycopg2-binary
